// src/app/pages/model/customer.ts
export class Customer {
     customerId: number;
     fullName: string;
     email: string;
     mobileNo: string;
     loginId: string;
     loginPassword: string;
    
     village: string;
     postOffice: string;
     policeStation: string;
     countryName: string;
     stateName: string;
     districtName: string;
    
     dateOfBirth: Date;
     gender: string;
     hobbies: any;
     education: string;
     comment: string;
    
     customerList: any;
     errorMsg: string;
    
     role: string;
     status: string;
     createdBy: number;
     createdOn: Date;
     modifiedBy: number;
     modifiedOn: Date;
 }